[Build] Make both Clang and GCC happy and reduce warnings (#22356)
it's hard to please both compilers, but this is an attempt

MODULAR_ORIG_COMMIT_REV_ID: 6692261b0d58fb00936b3f30dd60ea3a39476e07

+++ KGEN/lib/KGENToLLVM/LowerKGENClosuresToLLVM.cpp
@@ -255,7 +255,7 @@ private:
       Type boundArgTy = types.boundArgTypes[argIdx];
       Value boundArgPtr = rewriter.create<LLVM::GEPOp>(
           op.getLoc(), LLVM::LLVMPointerType::get(boundArgTy), envStruct,
-          ArrayRef<LLVM::GEPArg>({0, argIdx}));
+          ArrayRef<LLVM::GEPArg>({0, static_cast<int32_t>(argIdx)}));
       rewriter.create<LLVM::StoreOp>(op.getLoc(), boundArgValue, boundArgPtr);
     }
